Add process status management; implement ProcessRegistry for monitoring processes and enhance SystemProcess with status reporting
