+++ src/logger/console_logger.cpp
@@ -38,7 +38,8 @@ namespace tools
         }
 
     private:
-        BufferedConsoleLogger() : bufferLimit_(process::ProcessController::numProcesses() * 2)
+        //BufferedConsoleLogger() : bufferLimit_(process::ProcessController::numProcesses() * 2)
+        BufferedConsoleLogger() : bufferLimit_(1)
         {
         }
         std::vector<std::string> buffer_;
