+++ src/process/systemprocess/system_process.cpp
@@ -1,23 +1,28 @@
 #include "system_process.h"
-#include <cstring>
 #include "logger_instance.h"
-#include "random_stuff.h"
-#include <unistd.h>
+#include "communicator.h"
+#include "process_status.h"
+#include <thread>
+#include "nap_time.h"
 namespace process
 {
     void SystemProcess::work()
     {
-        preWork(getpid());
-        Arguments arguments;
-        while (proceed())
-        {
+        std::thread workerThread([this]() {
+            Arguments arguments;
             SpawnChild spawnChild(this, arguments.args);
-
-            tools::sleepRandomMs(true);
-            timeManager_.currentTime() = std::chrono::high_resolution_clock::now();
-        }
-        postWork();
+        });
+        workerThread.detach();
     }
+
+    void SystemProcess::postWork()
+    {
+        logLifetime();
+        if (getppid() != 1)
+            concurrency::Communicator::getInstance().sendTerminationMessage(timeManager_.getFormattedElapsedTimeStr(),
+                                                                            pid_, exitCode_);
+    }
+
 
     SystemProcess::SpawnChild::SpawnChild(SystemProcess *parent, const std::vector<char *> &args) : parent_(parent)
     {
@@ -32,6 +37,18 @@ namespace process
         {
             tools::LoggerManager::getInstance().logError("[SYSTEM PROCESS] Failed to execute command");
         }
+        else
+        {
+            parent_->preWork(parent_->pid_);
+            if (const pid_t result = waitpid(parent_->pid_, &status, 0); result != parent_->pid_)
+            {
+                tools::displayProcessStatus(status, parent_->pid_);
+            }
+            parent_->postWork();
+            std::this_thread::sleep_for(std::chrono::seconds(tools::NapTimeSec::MEDIUMS));
+
+        }
+
     }
 
     SystemProcess::SpawnChild::~SpawnChild()
